Support an explicit list of fixed-width types

+++ tests/test.cpp
@@ -354,13 +354,13 @@ TEST(move_semantics)
 TEST(floats)
 {
   Bytestream bts;
-  bts << (float64_t)0.0F << (float64_t)1 << (float64_t)1.1 << (float64_t)12.7
-      << 0.01171875 << 3.14159F;
+  bts << (float64_t)0.0 << (float64_t)1 << (float64_t)1.1 << (float64_t)12.7
+      << (float64_t)0.01171875 << (float32_t)3.14159;
   bts << 0x4037000000000000;
   bts.put<float64_t>((float64_t)666.777);
-  bts >> (float64_t)0.0F >> (float64_t)1 >> (float64_t)(1.1) >> (float64_t)12.7
-      >> (3.0/256) >> (float32_t)3.14159;
-  bts >> (float64_t)23.0 >> 666.777;
+  bts >> (float64_t)0.0 >> (float64_t)1 >> (float64_t)(1.1) >> (float64_t)12.7
+      >> (float64_t)(3.0/256) >> (float32_t)3.14159;
+  bts >> (float64_t)23.0 >> (float64_t)666.777;
 }
 
 TEST(endianness)
@@ -459,7 +459,7 @@ TEST(eject)
   ASSERT(bts == bts2);
 }
 
-#include "TestCodable.h"
+#include "testcodable.h"
 
 TEST(codable)
 {
@@ -501,7 +501,7 @@ TEST(codable)
   cod2.g = -3;
   cod2.h = -4;
   cod2.s = "someString";
-  cod2.f1 = 1.1;
+  cod2.f1 = (float32_t)1.1;
   cod2.f2 = -2.2;
   cod2.s2 = "short";
